Instances: Split default interface port connections
Splits up behavior for instantiating default interface port connections into a separate file.
This is so it's more clear which code is used in the abnormal / slang extensions flow.
I'm planning on adding $static_assert iface param pinning in a followup that requires additional code.

stack-info: PR: https://github.com/MikePopoloski/slang/pull/1909, branch: AndrewNolte/stack/23
